Draft proposal for simplified instance definition and implementations
The Instance interface is the base interface for all instances.
It clearly separates features (attributes) from labels (classes) and other properties.

Instance objects are immutable. This is not strictly enforced, but we normally
build/copy them through Builder classes, and there is no setter.

Default implementations are either dense or sparse.

The legacy code was tied to the ARFF format and assumed that a stream
- was bounded in length (see the "Instances" class)
- provided all metadata upfront (InstancesHeader)

We lift this assumption in this refactoring.

One thing which is missing (code does not compile) is the number of classes, required
in the various algorithms. In the legacy code, this was piggy-backed to each instance.
We need another way to pass that information.